Refactors generation code to avoid copying things into vecs over and over and over.
We now pass around a &mut Vec wrapped as an "OutputStream", which only permits insertion.

While this may seem less flexible, we can technically still have a vec in some functions if we need to, by creating a new one and building what is essentially a "sub OutputIterator". But we haven't need that yet.